+++ R/Tools4InsilicoMsMs.R
@@ -21,8 +21,8 @@
 setGeneric(name = 'convertAdduct4Tools',
            def = function(
              adduct,
-             tool_from = c('metdna2', 'msfinder', 'sirius'),
-             tool_to = c('metdna2', 'msfinder', 'sirius')
+             tool_from = c('metdna2', 'msfinder', 'sirius', 'metdna'),
+             tool_to = c('metdna2', 'msfinder', 'sirius', 'metdna'),
            ){
              tool_from <- match.arg(tool_from)
              tool_to <- match.arg(tool_to)
@@ -31,12 +31,14 @@ setGeneric(name = 'convertAdduct4Tools',
                      'metdna2' = {list_adduct_from <- lib_adduct_conv$MetDNA2},
                      'msfinder' = {list_adduct_from <- lib_adduct_conv$MSFINDER},
                      'sirius' = {list_adduct_from <- lib_adduct_conv$SIRIUS}
+                     'metdna' = {list_adduct_from <- lib_adduct_conv$MetDNA}
              )
 
              switch (tool_to,
                      'metdna2' = {list_adduct_to <- lib_adduct_conv$MetDNA2},
                      'msfinder' = {list_adduct_to <- lib_adduct_conv$MSFINDER},
-                     'sirius' = {list_adduct_to <- lib_adduct_conv$SIRIUS}
+                     'sirius' = {list_adduct_to <- lib_adduct_conv$SIRIUS},
+                     'sirius' = {list_adduct_to <- lib_adduct_conv$MetDNA}
              )
 
              idx <- match(adduct, list_adduct_from)